requirements.txt
torch>=2.3.0
transformers>=4.44.0
peft>=0.12.0
datasets>=2.20.0
accelerate>=0.33.0
scipy>=1.13.0
numpy>=1.26.0
pyyaml>=6.0
wandb>=0.17.0
tqdm>=4.66.0
